Add hook to load image
